+++ AltCover.PowerShell/Command.fs
@@ -144,6 +144,16 @@ type InvokeAltCoverCommand(runner : bool) =
               ValueFromPipeline = false, ValueFromPipelineByPropertyName = false)>]
   member val Version : SwitchParameter = SwitchParameter(false) with get, set
 
+  [<Parameter(ParameterSetName = "Instrument", Mandatory = false,
+              ValueFromPipeline = false, ValueFromPipelineByPropertyName = false)>]
+  [<Parameter(ParameterSetName = "Runner", Mandatory = false, ValueFromPipeline = false,
+              ValueFromPipelineByPropertyName = false)>]
+  member val DropReturnCode : SwitchParameter = SwitchParameter(false) with get, set
+
+  [<Parameter(ParameterSetName = "Instrument", Mandatory = false,
+              ValueFromPipeline = false, ValueFromPipelineByPropertyName = false)>]
+  member val SourceLink : SwitchParameter = SwitchParameter(false) with get, set
+
   member val private Fail : String list = [] with get, set
 
   member private self.Collect() =
@@ -154,7 +164,8 @@ type InvokeAltCoverCommand(runner : bool) =
                                     Threshold = self.Threshold
                                     Cobertura = self.Cobertura
                                     OutputFile = self.OutputFile
-                                    CommandLine = self.CommandLine }
+                                    CommandLine = self.CommandLine
+                                    ExposeReturnCode = not self.DropReturnCode.IsPresent }
 
   member private self.Prepare() =
     FSApi.PrepareParams.Primitive { InputDirectory = self.InputDirectory
@@ -184,7 +195,9 @@ type InvokeAltCoverCommand(runner : bool) =
                                     Single = self.Single.IsPresent
                                     LineCover = self.LineCover.IsPresent
                                     BranchCover = self.BranchCover.IsPresent
-                                    CommandLine = self.CommandLine }
+                                    CommandLine = self.CommandLine
+                                    ExposeReturnCode = not self.DropReturnCode.IsPresent
+                                    SourceLink = self.SourceLink.IsPresent }
 
   member private self.Log() =
     FSApi.Logging.Primitive { Primitive.Logging.Create() with Error = (fun s -> self.Fail <- s :: self.Fail)