+++ AltCover/AltCover.fs
@@ -21,6 +21,7 @@ type internal AssemblyInfo =
 module internal Main =
   let init() =
     CommandLine.error <- []
+    CommandLine.dropReturnCode := false
     Visitor.inputDirectory <- None
     Visitor.outputDirectory <- None
     ProgramDatabase.SymbolFolders.Clear()
@@ -41,10 +42,11 @@ module internal Main =
     Visitor.interval <- None
     Visitor.TrackingNames.Clear()
     Visitor.reportFormat <- None
-    Visitor.inplace <- false
-    Visitor.collect <- false
+    Visitor.inplace := false
+    Visitor.collect := false
     Visitor.single <- false
     Visitor.coverstyle <- CoverStyle.All
+    Visitor.sourcelink := false
 
   let ValidateCallContext predicate x =
     if not (String.IsNullOrWhiteSpace x) then
@@ -195,22 +197,8 @@ module internal Main =
                                  CommandLine.resources.GetString "MultiplesNotAllowed",
                                  "--opencover") :: CommandLine.error
        else Visitor.reportFormat <- Some ReportFormat.OpenCover))
-      ("inplace",
-       (fun _ ->
-       if Visitor.inplace then
-         CommandLine.error <- String.Format
-                                (CultureInfo.CurrentCulture,
-                                 CommandLine.resources.GetString "MultiplesNotAllowed",
-                                 "--inplace") :: CommandLine.error
-       else Visitor.inplace <- true))
-      ("save",
-       (fun _ ->
-       if Visitor.collect then
-         CommandLine.error <- String.Format
-                                (CultureInfo.CurrentCulture,
-                                 CommandLine.resources.GetString "MultiplesNotAllowed",
-                                 "--save") :: CommandLine.error
-       else Visitor.collect <- true))
+      (CommandLine.ddFlag "inplace" Visitor.inplace)
+      (CommandLine.ddFlag "save" Visitor.collect)
       ("single",
        (fun _ ->
        if Visitor.single then
@@ -252,6 +240,8 @@ module internal Main =
                                  CommandLine.resources.GetString "Incompatible",
                                  "--branchcover", "--linecover") :: CommandLine.error
        | _ -> Visitor.coverstyle <- CoverStyle.BranchOnly))
+      (CommandLine.ddFlag "dropReturnCode" CommandLine.dropReturnCode)
+      (CommandLine.ddFlag "sourcelink" Visitor.sourcelink)
       ("?|help|h", (fun x -> CommandLine.help <- not (isNull x)))
 
       ("<>",
@@ -275,7 +265,7 @@ module internal Main =
         CommandLine.error <- CommandLine.resources.GetString "NotInPlace"
                              :: CommandLine.error
       CommandLine.doPathOperation (fun () ->
-        if Visitor.inplace && CommandLine.error |> List.isEmpty
+        if !Visitor.inplace && CommandLine.error |> List.isEmpty
            && toDirectory |> Directory.Exists then
           CommandLine.error <- String.Format
                                  (CultureInfo.CurrentCulture,
@@ -288,7 +278,7 @@ module internal Main =
          |> not
       then Left("UsageError", options)
       else
-        if Visitor.inplace then
+        if !Visitor.inplace then
           Output.Info
           <| String.Format
                (CultureInfo.CurrentCulture, (CommandLine.resources.GetString "savingto"),
@@ -443,9 +433,9 @@ module internal Main =
 
           Visitor.Visit visitors (assemblies)
           document.Save(report)
-          if Visitor.collect then Runner.SetRecordToFile report
+          if !Visitor.collect then Runner.SetRecordToFile report
           CommandLine.ProcessTrailingArguments rest toInfo) 255 true
-      CommandLine.ReportErrors "Instrumentation" (dotnetBuild && Visitor.inplace)
+      CommandLine.ReportErrors "Instrumentation" (dotnetBuild && !Visitor.inplace)
       result
 
   let internal (|Select|_|) (pattern : String) offered =